moved to preprocessor
