+++ src/Converter/ConfluenceConverter.php
@@ -20,6 +20,7 @@ use HalloWelt\MigrateConfluence\Converter\Postprocessor\NestedHeadings;
 use HalloWelt\MigrateConfluence\Converter\Postprocessor\RestorePStyleTag;
 use HalloWelt\MigrateConfluence\Converter\Postprocessor\RestoreTimeTag;
 use HalloWelt\MigrateConfluence\Converter\Postprocessor\TasksReportMacro as RestoreTasksReportMacro;
+use HalloWelt\MigrateConfluence\Converter\Preprocessor\dom\HoistMacroFromHeading;
 use HalloWelt\MigrateConfluence\Converter\Preprocessor\dom\SanitizeLinkContent;
 use HalloWelt\MigrateConfluence\Converter\Preprocessor\html\CDATAClosingFixer;
 use HalloWelt\MigrateConfluence\Converter\Processor\AlignMacro;
@@ -40,7 +41,6 @@ use HalloWelt\MigrateConfluence\Converter\Processor\ExcerptMacro;
 use HalloWelt\MigrateConfluence\Converter\Processor\ExpandMacro;
 use HalloWelt\MigrateConfluence\Converter\Processor\GalleryMacro;
 use HalloWelt\MigrateConfluence\Converter\Processor\GliffyMacro;
-use HalloWelt\MigrateConfluence\Converter\Processor\HoistMacroFromHeading;
 use HalloWelt\MigrateConfluence\Converter\Processor\Image;
 use HalloWelt\MigrateConfluence\Converter\Processor\IncludeMacro;
 use HalloWelt\MigrateConfluence\Converter\Processor\InfoMacro;
@@ -311,7 +311,6 @@ class ConfluenceConverter extends PandocHTML implements IOutputAwareInterface {
 			new Layout(),
 			new LayoutSection(),
 			new LayoutCell(),
-			new HoistMacroFromHeading(),
 			new AnchorMacro(),
 			new Placeholder(),
 			new InlineCommentMarker(),
@@ -579,10 +578,11 @@ class ConfluenceConverter extends PandocHTML implements IOutputAwareInterface {
 	 */
 	protected function preprocessDomSource( DOMDocument $dom ): void {
 		$preprocessors = [
-			new SanitizeLinkContent()
+			new SanitizeLinkContent(),
+			new HoistMacroFromHeading()
 		];
 
-		/** @var IHtmlPreprocessor $preprocessor */
+		/** @var IDomPreprocessor $preprocessor */
 		foreach ( $preprocessors as $preprocessor ) {
 			$preprocessor->preprocess( $dom );
 		}
